pyproject.toml
[tool.poetry]
name = "weatherapp"
version = "0.1.0"
description = ""
authors = ["Ivan <ivan4dante@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests-async = "^0.6.2"
flask = {extras = ["async"], version = "^3.0.3"}
peewee = "^3.17.6"
peewee-aio = "^1.7.4"
pyjwt = "^2.8.0"
pytest = "^8.2.2"
pytest-env = "^1.1.3"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.23.8"

[tool.pytest.ini_options]
addopts = "-xslv"
python_files = "test*.py"
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::pytest.PytestUnraisableExceptionWarning"
]
env = [
    "ENV=tests"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
